Refactor: 컴포넌트 명 오타 제거 (MM-44)
